Move DataHolder to keep organization.
