+++ src/main/java/org/spongepowered/api/block/BlockLoc.java
@@ -27,7 +27,7 @@ package org.spongepowered.api.block;
 
 import com.flowpowered.math.vector.Vector3i;
 import org.spongepowered.api.item.inventory.ItemStack;
-import org.spongepowered.api.util.DataHolder;
+import org.spongepowered.api.service.persistence.data.DataHolder;
 import org.spongepowered.api.util.Direction;
 import org.spongepowered.api.world.Location;
 import org.spongepowered.api.world.extent.Extent;
