+++ src/main/java/org/spongepowered/api/entity/Entity.java
@@ -27,7 +27,7 @@ package org.spongepowered.api.entity;
 
 import com.flowpowered.math.vector.Vector2f;
 import com.google.common.base.Optional;
-import org.spongepowered.api.util.DataHolder;
+import org.spongepowered.api.service.persistence.data.DataHolder;
 import org.spongepowered.api.util.Identifiable;
 import org.spongepowered.api.world.Location;
 import org.spongepowered.api.world.World;
